backend/requirements.txt
pandas
matplotlib
python-dotenv
tqdm
psycopg2-binary
langchain-google-genai
streamlit
